Double click detection added + item equipment fixes
- Updated actor's status bar display.
- Fixed multiple Infinite fall coroutine calls.
- Remade Gear equipment method.
- Item pickup via mouse click on it while in inventory.
- Implemented new DoubleClick detection static class.

+++ Assets/Scripts/Door.cs
@@ -63,7 +63,7 @@ public class Door : MonoBehaviour, IInteractable, IDamageable
         }
     }
 
-    public void GetHit(int amount, DamageType type, Transform part)
+    public void GetHit(float amount, DamageType type, Transform part)
     {
         SM.PlaySound("GetHit");
