+++ Assets/Scripts/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void GetHit(int damage, DamageType type, Transform part);
+    public void GetHit(float damage, DamageType type, Transform part);
 }
